Made slight code improvements over the weekend since I brought an iPad instead of a laptop to the workshop

+++ app/src/main/java/example/Calculator.java
@@ -4,17 +4,13 @@ public class Calculator {
     public int add(int a, int b) {
         return a + b;
     }
-
     public int minus(int a, int b) {
         return a - b;
     }
-
     public int multiply(int a, int b) {
-
-            return a*b;
+        return a*b;
     }
-
     public double divide(int a, int b) {
-        return a/b;
+        return (double) a/b;
     }
 }
